Program/requirements.txt
flask
azure-data-tables
azure-eventhub